refactor(trpc/routes): change gameplay routes over to new file structure and rbac
