+++ apps/app/server/trpc/rbac.ts
@@ -1,0 +1,12 @@
+/**
+ * 
+ * @param userScope 
+ * The current logged on users scope;
+ * @param requiredScope 
+ * The required scope
+ * @returns boolean
+ */
+export const userHasScope = (userScope: Array<string>, requiredScope: Array<string>): boolean => {
+    const userHasRequiredScope = requiredScope.some(a => userScope.includes(a));
+    return userHasRequiredScope;
+}
